+++ src/sasktran2/optical/__init__.py
@@ -4,7 +4,7 @@ import sasktran2 as sk
 from sasktran2.constants import K_BOLTZMANN
 from sasktran2.database.web import StandardDatabase
 
-from . import database
+from . import database, refraction  # noqa: F401
 
 
 class O3DBM(database.OpticalDatabaseGenericAbsorber):